+++ src/main/java/woowacourse/shoppingcart/controller/CartItemController.java
@@ -3,52 +3,64 @@ package woowacourse.shoppingcart.controller;
 import java.net.URI;
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import woowacourse.shoppingcart.domain.cart.Cart;
-import woowacourse.shoppingcart.domain.product.Product;
-import woowacourse.shoppingcart.dto.Request;
-import woowacourse.shoppingcart.service.CartService;
+import woowacourse.auth.support.AuthenticationPrincipal;
+import woowacourse.shoppingcart.dto.CartItemDeletionRequest;
+import woowacourse.shoppingcart.dto.CartItemRequest;
+import woowacourse.shoppingcart.dto.CartItemResponse;
+import woowacourse.shoppingcart.dto.UpdateCartItemRequest;
+import woowacourse.shoppingcart.service.CartItemService;
 
 @RestController
-@RequestMapping("/api/customers/{customerName}/carts")
+@RequestMapping("/api/mycarts")
 public class CartItemController {
-    private final CartService cartService;
 
-    public CartItemController(final CartService cartService) {
-        this.cartService = cartService;
-    }
+    private final CartItemService cartItemService;
 
-    @GetMapping
-    public ResponseEntity<List<Cart>> getCartItems(@PathVariable final String customerName) {
-        return ResponseEntity.ok().body(cartService.findCartsByCustomerName(customerName));
+    public CartItemController(CartItemService cartItemService) {
+        this.cartItemService = cartItemService;
     }
 
     @PostMapping
-    public ResponseEntity<Void> addCartItem(@Validated(Request.id.class) @RequestBody final Product product,
-        @PathVariable final String customerName) {
-        final Long cartId = cartService.addCart(product.getId(), customerName);
-        final URI responseLocation = ServletUriComponentsBuilder
-            .fromCurrentRequest()
-            .path("/{cartId}")
-            .buildAndExpand(cartId)
-            .toUri();
-        return ResponseEntity.created(responseLocation).build();
+    public ResponseEntity<CartItemResponse> addCart(@AuthenticationPrincipal String email,
+        @Valid @RequestBody CartItemRequest cartItemRequest) {
+        CartItemResponse cartItemResponse = cartItemService.addCart(email, cartItemRequest);
+        return ResponseEntity.created(URI.create("/api/mycarts/" + cartItemResponse.getId())).body(cartItemResponse);
     }
 
-    @DeleteMapping("/{cartId}")
-    public ResponseEntity<Void> deleteCartItem(@PathVariable final String customerName,
-        @PathVariable final Long cartId) {
-        cartService.deleteCart(customerName, cartId);
+    @GetMapping
+    public ResponseEntity<List<CartItemResponse>> findCartItems(@AuthenticationPrincipal String email) {
+        return ResponseEntity.ok(cartItemService.getCartItems(email));
+    }
+
+    @GetMapping("/{cartItemId}")
+    public ResponseEntity<CartItemResponse> findCartItem(@AuthenticationPrincipal String email,
+        @PathVariable(value = "cartItemId") long id) {
+        return ResponseEntity.ok(cartItemService.getCartItem(email, id));
+    }
+
+    @PatchMapping
+    public ResponseEntity<Void> updateQuantity(@AuthenticationPrincipal String email,
+        @RequestBody UpdateCartItemRequest cartItemRequest) {
+        cartItemService.update(email, cartItemRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping
+    public ResponseEntity<Void> delete(@AuthenticationPrincipal String email,
+        @RequestBody CartItemDeletionRequest cartItemDeletionRequest) {
+        cartItemService.delete(email, cartItemDeletionRequest);
         return ResponseEntity.noContent().build();
     }
 }
