+++ src/main/java/woowacourse/shoppingcart/dao/CartItemDao.java
@@ -29,40 +29,6 @@ public class CartItemDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public List<Long> findProductIdsByCustomerId(final Long customerId) {
-        final String query = "SELECT product_id FROM cart_item WHERE customer_id = ?";
-
-        return jdbcTemplate.query(query, (rs, rowNum) -> rs.getLong("product_id"), customerId);
-    }
-
-    public List<Long> findIdsByCustomerId(final Long customerId) {
-        final String query = "SELECT id FROM cart_item WHERE customer_id = ?";
-
-        return jdbcTemplate.query(query, (rs, rowNum) -> rs.getLong("id"), customerId);
-    }
-
-    public Long findProductIdById(final Long cartId) {
-        try {
-            final String query = "SELECT product_id FROM cart_item WHERE id = ?";
-            return jdbcTemplate.queryForObject(query, (rs, rowNum) -> rs.getLong("product_id"), cartId);
-        } catch (EmptyResultDataAccessException e) {
-            throw new InvalidCartItemException();
-        }
-    }
-
-    public Long addCartItem(final long customerId, final long productId) {
-        final String query = "INSERT INTO cart_item(customer_id, product_id) VALUES(?, ?)";
-        final KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(con -> {
-            PreparedStatement preparedStatement = con.prepareStatement(query, new String[] {"id"});
-            preparedStatement.setLong(1, customerId);
-            preparedStatement.setLong(2, productId);
-            return preparedStatement;
-        }, keyHolder);
-        return keyHolder.getKey().longValue();
-    }
-
     public long addCartItem(final long customerId, final CartItem cartItem) {
         final String query = "INSERT INTO cart_item(customer_id, product_id, quantity) VALUES(?, ?, ?)";
         final KeyHolder keyHolder = new GeneratedKeyHolder();
@@ -76,15 +42,6 @@ public class CartItemDao {
         }, keyHolder);
 
         return keyHolder.getKey().longValue();
-    }
-
-    public void deleteCartItem(final Long id) {
-        final String sql = "DELETE FROM cart_item WHERE id = ?";
-
-        final int rowCount = jdbcTemplate.update(sql, id);
-        if (rowCount == 0) {
-            throw new InvalidCartItemException();
-        }
     }
 
     public List<CartItemEntity> findCartItemsByCustomerId(long customerId) {
