+++ src/main/java/woowacourse/shoppingcart/dao/OrdersDetailDao.java
@@ -9,7 +9,6 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
-import woowacourse.shoppingcart.domain.NewOrderDetail;
 import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.entity.OrderDetailEntity;
 
@@ -27,34 +26,12 @@ public class OrdersDetailDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Long addOrdersDetail(final Long ordersId, final Long productId, final int quantity) {
-        final String query = "INSERT INTO orders_detail (orders_id, product_id, quantity) VALUES (?, ?, ?)";
-        final KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(con -> {
-            PreparedStatement preparedStatement = con.prepareStatement(query, new String[] {"id"});
-            preparedStatement.setLong(1, ordersId);
-            preparedStatement.setLong(2, productId);
-            preparedStatement.setLong(3, quantity);
-            return preparedStatement;
-        }, keyHolder);
-        return keyHolder.getKey().longValue();
-    }
-
-    public List<OrderDetail> findOrdersDetailsByOrderId(final Long orderId) {
-        final String query = "SELECT product_id, quantity FROM orders_detail WHERE orders_id = ?";
-        return jdbcTemplate.query(query, (rs, rowNum) -> new OrderDetail(
-            rs.getLong("product_id"),
-            rs.getInt("quantity")
-        ), orderId);
-    }
-
     public List<OrderDetailEntity> findOrderDetailsByOrderId(long ordersId) {
         final String sql = "SELECT orders_id, product_id, quantity FROM orders_detail WHERE orders_id = ?";
         return jdbcTemplate.query(sql, ORDER_DETAIL_ENTITY_ROW_MAPPER, ordersId);
     }
 
-    public Long add(Long orderId, NewOrderDetail orderDetail) {
+    public Long add(Long orderId, OrderDetail orderDetail) {
         final String query = "INSERT INTO orders_detail(orders_id, product_id, quantity) VALUES(?, ?, ?)";
         final KeyHolder keyHolder = new GeneratedKeyHolder();
