--- src/main/java/woowacourse/shoppingcart/domain/NewOrderDetail.java
@@ -1,25 +1,0 @@
-package woowacourse.shoppingcart.domain;
-
-import woowacourse.shoppingcart.domain.product.Product;
-
-public class NewOrderDetail {
-    private final Product product;
-    private final Quantity quantity;
-
-    public NewOrderDetail(Product product, Quantity quantity) {
-        this.product = product;
-        this.quantity = quantity;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public long getProductId() {
-        return product.getId();
-    }
-
-    public int getQuantity() {
-        return quantity.getQuantity();
-    }
-}
