+++ src/main/java/woowacourse/shoppingcart/domain/OrderDetail.java
@@ -3,50 +3,23 @@ package woowacourse.shoppingcart.domain;
 import woowacourse.shoppingcart.domain.product.Product;
 
 public class OrderDetail {
-    private long productId;
-    private int quantity;
-    private int price;
-    private String name;
-    private String imageUrl;
+    private final Product product;
+    private final Quantity quantity;
 
-    public OrderDetail() {
-    }
-
-    public OrderDetail(final long productId, final int quantity) {
-        this.productId = productId;
+    public OrderDetail(Product product, Quantity quantity) {
+        this.product = product;
         this.quantity = quantity;
     }
 
-    public OrderDetail(final Product product, final int quantity) {
-        this(product.getId(), product.getPrice(), product.getName(), product.getThumbnailImageUrl(), quantity);
-    }
-
-    public OrderDetail(final Long productId, final int price, final String name,
-        final String imageUrl, final int quantity) {
-        this.productId = productId;
-        this.price = price;
-        this.name = name;
-        this.imageUrl = imageUrl;
-        this.quantity = quantity;
+    public Product getProduct() {
+        return product;
     }
 
     public long getProductId() {
-        return productId;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
+        return product.getId();
     }
 
     public int getQuantity() {
-        return quantity;
+        return quantity.getQuantity();
     }
 }
