+++ src/main/java/woowacourse/shoppingcart/domain/Orders.java
@@ -4,10 +4,15 @@ import java.util.List;
 
 public class Orders {
 
-    private final Long id;
+    private Long id;
+
     private final List<OrderDetail> orderDetails;
 
-    public Orders(final Long id, final List<OrderDetail> orderDetails) {
+    public Orders(List<OrderDetail> orderDetails) {
+        this(null, orderDetails);
+    }
+
+    public Orders(Long id, List<OrderDetail> orderDetails) {
         this.id = id;
         this.orderDetails = orderDetails;
     }
