--- src/main/java/woowacourse/shoppingcart/dto/DeleteCartItemRequest.java
@@ -1,18 +1,0 @@
-package woowacourse.shoppingcart.dto;
-
-import java.util.List;
-
-public class DeleteCartItemRequest {
-    private List<Long> cartItemIds;
-
-    public DeleteCartItemRequest() {
-    }
-
-    public DeleteCartItemRequest(List<Long> cartItemIds) {
-        this.cartItemIds = cartItemIds;
-    }
-
-    public List<Long> getCartItemIds() {
-        return cartItemIds;
-    }
-}
