+++ src/main/java/woowacourse/shoppingcart/dto/OrderRequest.java
@@ -1,24 +1,18 @@
 package woowacourse.shoppingcart.dto;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
+import java.util.List;
 
 public class OrderRequest {
-    @NotNull
-    private final Long cartId;
-    @Min(0)
-    private final int quantity;
+    private List<Long> cartItemIds;
 
-    public OrderRequest(final Long cartId, final int quantity) {
-        this.cartId = cartId;
-        this.quantity = quantity;
+    public OrderRequest() {
     }
 
-    public Long getCartId() {
-        return cartId;
+    public OrderRequest(List<Long> cartItemIds) {
+        this.cartItemIds = cartItemIds;
     }
 
-    public int getQuantity() {
-        return quantity;
+    public List<Long> getCartItemIds() {
+        return cartItemIds;
     }
 }
