+++ src/main/java/woowacourse/shoppingcart/dto/OrderResponse.java
@@ -3,7 +3,7 @@ package woowacourse.shoppingcart.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import woowacourse.shoppingcart.domain.NewOrders;
+import woowacourse.shoppingcart.domain.Orders;
 
 public class OrderResponse {
     private final long id;
@@ -14,7 +14,7 @@ public class OrderResponse {
         this.orderedProducts = orderedProducts;
     }
 
-    public static OrderResponse from(NewOrders orders) {
+    public static OrderResponse from(Orders orders) {
         return new OrderResponse(orders.getId(), orders.getOrderDetails().stream()
             .map(OrderedProductDto::from)
             .collect(Collectors.toList()));
