+++ src/test/java/woowacourse/shoppingcart/dao/OrdersDetailDaoTest.java
@@ -1,20 +1,12 @@
 package woowacourse.shoppingcart.dao;
 
-import static org.assertj.core.api.Assertions.*;
-
-import java.util.List;
-
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
-
-import woowacourse.shoppingcart.domain.OrderDetail;
 
 @JdbcTest
 @AutoConfigureTestDatabase(replace = Replace.NONE)
@@ -43,38 +35,5 @@ class OrdersDetailDaoTest {
             "INSERT INTO product (name, price, stock_quantity, thumbnail_url, thumbnail_alt) VALUES (?, ?, ?, ?, ?)"
             , "name", 1000, 10, "imageUrl", "imageAlt");
         productId = jdbcTemplate.queryForObject("SELECT LAST_INSERT_ID();", Long.class);
-    }
-
-    @DisplayName("OrderDatail을 추가하는 기능")
-    @Test
-    void addOrdersDetail() {
-        //given
-        int quantity = 5;
-
-        //when
-        Long orderDetailId = ordersDetailDao
-            .addOrdersDetail(ordersId, productId, quantity);
-
-        //then
-        assertThat(orderDetailId).isEqualTo(1L);
-    }
-
-    @DisplayName("OrderId로 OrderDetails 조회하는 기능")
-    @Test
-    void findOrdersDetailsByOrderId() {
-        //given
-        final int insertCount = 3;
-        for (int i = 0; i < insertCount; i++) {
-            jdbcTemplate
-                .update("INSERT INTO orders_detail (orders_id, product_id, quantity) VALUES (?, ?, ?)",
-                    ordersId, productId, 3);
-        }
-
-        //when
-        final List<OrderDetail> ordersDetailsByOrderId = ordersDetailDao
-            .findOrdersDetailsByOrderId(ordersId);
-
-        //then
-        assertThat(ordersDetailsByOrderId).hasSize(insertCount);
     }
 }
